Finish
Tests ok. Some modifications may be done later, maybe...
