+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/JavaCXFCodegen.java
@@ -1,0 +1,73 @@
+
+package io.swagger.codegen.languages;
+
+import java.io.File;
+import java.util.List;
+import java.util.Map;
+
+import io.swagger.codegen.CliOption;
+import io.swagger.codegen.CodegenConstants;
+import io.swagger.codegen.CodegenOperation;
+import io.swagger.models.Operation;
+
+public class JavaCXFCodegen extends AbstractJavaJAXRSServerCodegen
+{
+	public JavaCXFCodegen()
+	{
+        super();
+
+        sourceFolder = "src/gen/java";
+        invokerPackage = "io.swagger.api";
+        artifactId = "swagger-jaxrs-server";
+        outputFolder = "generated-code/JavaJaxRS-CXF";
+
+        modelTemplateFiles.put("model.mustache", ".java");
+        apiTemplateFiles.put("api.mustache", ".java");
+        apiPackage = "io.swagger.api";
+        modelPackage = "io.swagger.model";
+
+        additionalProperties.put("title", title);
+
+        super.embeddedTemplateDir = templateDir = JAXRS_TEMPLATE_DIRECTORY_NAME + File.separator + "cxf";
+
+        for ( int i = 0; i < cliOptions.size(); i++ ) {
+            if ( CodegenConstants.LIBRARY.equals(cliOptions.get(i).getOpt()) ) {
+                cliOptions.remove(i);
+                break;
+            }
+        }
+        
+        cliOptions.add(new CliOption(CodegenConstants.IMPL_FOLDER, CodegenConstants.IMPL_FOLDER_DESC));
+	}
+
+	@Override
+	public void processOpts()
+	{
+		super.processOpts();
+		sourceFolder = "gen" + File.separator + "java";
+
+		modelTemplateFiles.clear();
+		modelTemplateFiles.put("entityModel.mustache", ".java");
+
+		supportingFiles.clear();
+	} 
+
+	@Override
+	public String getName()
+	{
+		return "cxf";
+	}
+	
+
+    @Override
+    public void addOperationToGroup(String tag, String resourcePath, Operation operation, CodegenOperation co, Map<String, List<CodegenOperation>> operations) {
+        super.addOperationToGroup(tag, resourcePath, operation, co, operations);        
+        co.subresourceOperation = !co.path.isEmpty();
+    }
+
+    @Override
+    public String getHelp()
+    {
+        return "Generates a Java JAXRS Server application based on Apache CXF framework.";
+    }
+}
